Use longs inside SNPHWE c function
This prevents errors with sample sizes > ~300000, otherwise it would hit
an overflow error as ints wrap around

+++ snphwe/hwe.pyx
@@ -2,8 +2,7 @@
 # distutils: include_dirs = src/
 
 cdef extern from "snp_hwe.h":
-    # double SNPHWE(int, int, int) except +
-    double SNPHWE(int, int, int)
+    double SNPHWE(long, long, long)
 
 def snphwe(obs_hets, obs_hom1, obs_hom2):
     return SNPHWE(obs_hets, obs_hom1, obs_hom2)
